==> src/04_networkanalysis.Rmd <==
---
title: "Analyse result networks"
output:
  html_document:
    keep_md: yes
    toc: yes
    theme: united
    code_folding: hide
---

This script summarizes efforts performed to analyze the role of TPP proteins in COSMOS networks. We saw that the processes in the overall network reflect our expectations (DDR). We then move on to a more granular view, by analyzing processes on cluster level.

# General settings


```{r setup}
knitr::opts_chunk$set(
  echo = TRUE,
  warning = FALSE,
  message = FALSE,
  include = TRUE,
  cache = TRUE,
  cache.lazy = FALSE,
  eval = TRUE,
  fig.width = 7 * (1 + sqrt(5)) / 2, fig.height = 7,
  dpi = 700
)
knitr::opts_knit$set(root.dir = "C:/Users/burtsche/Documents/COSMOS-TPP_paper")
```

## Packages

```{r, message=F, warning =F}
library(tidyverse)
library(igraph)
library("reshape2")
library(OmnipathR)
library(viridis)
library(ggrepel)
options(connectionObserver = NULL)
library(org.Hs.eg.db)
library(visNetwork)
library(knitr)
library(ggpubr)
library(Rcpp)
library(RColorBrewer)
library(ReactomePA)
library(clusterProfiler)
library(ggraph)
library(reactome.db)
library(ggvenn)

mutate <- dplyr::mutate
select <- dplyr::select
group_by <- dplyr::group_by
```

## Functions

Function 1: extract and annotate relevant dfs of COSMOS rsult object for plotting and analysis.

```{r}
network_processing <- function(COSMOS_object_fw2, # result of the first run
                               COSMOS_object_rev2, # result of the second run
                               TPP_input, # TPP proteins used as input
                               dorothea, # TF-target interactions
                               KSN, # Kinase-target interactions
                               all_input_nodes){ # all input nodes
  TFs <- dorothea$source_genesymbol
  Kinases <- KSN$enzyme_genesymbol
  # get all nodes and convert to gene symbols
  fw_nodes <- COSMOS_object_fw2$nodeAtt %>%
  mutate(EntrezID = gsub("X", "", Node))
  rev_nodes <- COSMOS_object_rev2$nodeAtt %>%
  mutate(EntrezID = gsub("X", "", Node))
  all_nodes_entrezid <- unique(c(fw_nodes$EntrezID, rev_nodes$EntrezID))
  convertIDs <-  mapIds(org.Hs.eg.db, all_nodes_entrezid, 'SYMBOL', 'ENTREZID') 
  # get networks, map to gene symbols and combine
  fw_network <- COSMOS_object_fw2$weightedSIF
  fw_network_mapped <- data.frame(from = convertIDs[gsub("X","",fw_network$Node1)],
                                  sign = fw_network$Sign,
                                  to = convertIDs[gsub("X","",fw_network$Node2)]) %>%
                        mutate(smooth = TRUE,
                               dashes = ifelse(sign > 0, FALSE, TRUE))  %>%
                        distinct(from, sign, to, .keep_all = T)
                        # filter out big artificial cluster
                        # filter(from != "TP63")
  rev_network <- COSMOS_object_rev2$weightedSIF
  rev_network_mapped <- data.frame(from = convertIDs[gsub("X","",rev_network$Node1)],
                                  sign = rev_network$Sign,
                                  to = convertIDs[gsub("X","",rev_network$Node2)]) %>%
    mutate(smooth = TRUE,
           dashes = ifelse(sign > 0, FALSE, TRUE)) %>%
    distinct(from, sign, to, .keep_all = T)
  
  combined_network <- rbind(fw_network_mapped, rev_network_mapped)
  # extract activity information
  all_nodes <- rbind(filter(fw_nodes, AvgAct != 0), filter(rev_nodes, AvgAct != 0))%>%
  mutate(activity = ifelse(AvgAct > 0, "active", "inactive"))
  node_activities <- all_nodes$activity
  names(node_activities) <- mapIds(org.Hs.eg.db, all_nodes$EntrezID, 'SYMBOL', 'ENTREZID')
  missing_nodes <- setdiff(unique(c(combined_network$from, combined_network$to)), names(node_activities))
  dummy <- rep("not available", length(missing_nodes))
  names(dummy) <- missing_nodes
  node_activities <- c(node_activities, dummy)
  # map all information to nodes table
  mapped_nodes <- data.frame(id = unique(c(combined_network$from, combined_network$to)),
                             label = unique(c(combined_network$from, combined_network$to))) %>%
                  rowwise() %>%
                  mutate(measured = node_activities[[id]]) 
  mapped_nodes <- mapped_nodes %>%
    mutate(type = ifelse(id %in% Kinases, "Kinase", "other"),
           type = ifelse(id %in% TFs, "TF", type),
           type = ifelse(id %in% TPP_input, "TPP", type),
           shape = ifelse(id %in% all_input_nodes, "square", "triangle"),
           group = paste(type, measured, sep = "_"),
           title = paste0(label, "_", group))
  # store into visNetwork input object
  meta_data <- c(COSMOS_object_fw2$history, COSMOS_object_rev2$history)
  visNetwork_object <- structure(list(fw_network = fw_network_mapped,
                                      #fw_background = COSMOS_object_fw2$meta_network,
                                      rev_network = rev_network_mapped,
                                      #rev_background = COSMOS_object_rev2$meta_network,
                                      combined_network = combined_network,
                                      nodes = mapped_nodes,
                                      metadata = meta_data,
                                      class = "visNetworkinput"))
  
}
```

Function 2: extract main graph features of COSMOS networks

```{r}
network_checks <- function(visNetwork_object, # result of function 1
                           TPP_input, 
                           transcriptomic_regulators, 
                           phosphoproteomic_regulators, 
                           dorothea, 
                           KSN){
  # number of nodes and edges
  no_edges <- nrow(as.data.frame(visNetwork_object$combined_network))
  no_nodes <- nrow(as.data.frame(visNetwork_object$nodes))
  # convert to igraph graph object
  graph <-  graph_from_data_frame(as.data.frame(visNetwork_object$combined_network)[,c(1,3)], directed = TRUE, vertices = NULL)
  # edge density
  network_density <- edge_density(graph, loops = TRUE)
  # node degree centrality
  node_degrees <- degree(graph, v = V(graph), mode = "total",loops = TRUE, normalized = FALSE)
  node_degrees <- data.frame(Node = names(node_degrees), degree = node_degrees)
  # adjacency matrix
  adj_matrix <- as_adjacency_matrix(graph, type ="both",attr = NULL, edges = FALSE, names = TRUE)
  adj_matrix_m <- as.matrix(adj_matrix)
  adj_matrix_m[adj_matrix_m == 2] <- 1
  # all PKN nodes mapped to gene name identifiers
  # PKN_nodes <- unique(c(visNetwork_object$fw_background[["source"]], visNetwork_object$fw_background[["target"]],
  #                visNetwork_object$rev_background[["source"]], visNetwork_object$rev_background[["target"]]))
  # PKN_nodes_mapped <- mapIds(org.Hs.eg.db, gsub("X", "", PKN_nodes), 'SYMBOL', 'ENTREZID')
  # get input modeled, background modeled and missed input
  # TPP
  included_TPPhits <- intersect(visNetwork_object$nodes[["id"]], TPP_input)
  missedTPPhits <- setdiff(TPP_input, visNetwork_object$nodes[["id"]])
  #filtered_TPP <- setdiff(TPP_input, PKN_nodes_mapped)
  # TF
  included_TFs <- intersect(visNetwork_object$nodes[["id"]], transcriptomic_regulators)
  missedTFs <- setdiff(transcriptomic_regulators, visNetwork_object$nodes[["id"]])
  #filtered_TFs <- setdiff(transcriptomic_regulators, PKN_nodes_mapped)
  # Kinases
  included_kinases <- intersect(visNetwork_object$nodes[["id"]], phosphoproteomic_regulators)
  missedkinases <- setdiff(phosphoproteomic_regulators, visNetwork_object$nodes[["id"]])
  #filtered_kinases <- setdiff(phosphoproteomic_regulators, PKN_nodes_mapped)
  # summary
  input_nodes <- as.data.frame(visNetwork_object$nodes) %>% filter(shape == "square")
  input_nodes <- input_nodes$id
  additional_nodes <- as.data.frame(visNetwork_object$nodes) %>% filter(shape == "triangle")
  additional_nodes <- additional_nodes$id
  missed_nodes <- c(missedTPPhits, missedkinases, missedTFs)
  summary = data.frame(node= c(input_nodes, additional_nodes, missed_nodes),
                       model = c(rep("input", length(input_nodes)),
                                  rep("additional", length(additional_nodes)),
                                  rep("missed", length(missed_nodes)))) %>%
    mutate( type = ifelse(node %in% KSN$enzyme_genesymbol, "Kinase",  "other"),
            type = ifelse(node %in% dorothea$source_genesymbol, "TF",type),
            type = ifelse(node %in% TPP_input, "TPP", type))
  # build result object
  result <- structure(list(no_edges = no_edges,
                            no_nodes = no_nodes,
                            graph = graph,
                            adj_matrix = adj_matrix,
                            adj_matrix_m = adj_matrix_m,
                            node_degrees =node_degrees,
                            network_density = network_density,
                            overall_node_summary = summary))

}
```

Enrichment functions for GO and RPA enrichments

```{r}
GSE_RPA <- function(geneList, universe) {
  geneList <- mapIds(org.Hs.eg.db, geneList, "ENTREZID", "SYMBOL")
  universe <- mapIds(org.Hs.eg.db, universe, "ENTREZID", "SYMBOL")
  pathway_enrichment <- ReactomePA::enrichPathway(
    gene = geneList,
    organism = "human",
    pvalueCutoff = 1,
    pAdjustMethod = "BH",
    qvalueCutoff = 1,
    universe = universe,
    minGSSize = 5,
    maxGSSize = 500,
    readable = TRUE
  )
}

GSE_MF <- function(geneList, universe, ont) {
  geneList <- mapIds(org.Hs.eg.db, geneList, "ENTREZID", "SYMBOL")
  universe <- mapIds(org.Hs.eg.db, universe, "ENTREZID", "SYMBOL")
  pathway_enrichment <- clusterProfiler::enrichGO(
    gene = geneList,
    ont = ont,
    OrgDb = "org.Hs.eg.db",
    pvalueCutoff = 1,
    pAdjustMethod = "BH",
    qvalueCutoff = 1,
    universe = universe,
    minGSSize = 5,
    maxGSSize = 500,
    readable = TRUE
  )
}
```

## PKN data

```{r}
load("data/KSN.RData")
load("data/dorothea.RData")

load("results/220512_cosmosinput_merged.RData")

# to create universe for network level enrichment
PKN_universe <- read_tsv("data/2207_PKN_universe.tsv")
```

## Viper results

```{r}
load("data/220510_viper_footprints.RData")
viper_kinases <- viperRes_phospho_corrected %>%
  arrange(desc(absNES)) %>%
  slice_head(n = 30) %>%
  #adjust format to network nomenclature
  mutate(measured = ifelse(NES < 0, "inactive", "active"), type = "Kinase", node = regulator) %>%
  select(id = regulator, node, measured, type)


viper_TFs <- viperRes_trans %>%
  arrange(desc(absNES)) %>%
  slice_head(n = 30) %>%
  #adjust format to network nomenclature
  mutate(measured = ifelse(NES < 0, "inactive", "active"), type = "Kinase", node = regulator) %>%
  select(id = regulator, node, measured, type)
```

## Reactome terms

Extract reactome information for processes of interest.

```{r}
# fetch reactome data
pathwaymembers <- toTable(reactomePATHID2EXTID) %>%
  mutate(gene_name = mapIds(org.Hs.eg.db, gene_id, "SYMBOL", "ENTREZID"))
pathway_names <- toTable(reactomePATHID2NAME) %>%
  filter(grepl("Homo sapiens", path_name))
# extract information to pathways of interest
DDR <- pathway_names %>%
  filter(grepl("DNA Damage|DNA damage|DNA Repair|DNA repair|Checkpoint", path_name))
DDR_members <- pathwaymembers %>%
  filter(DB_ID %in% DDR$DB_ID) %>%
  distinct(gene_name)
interferon <- pathway_names %>%
  filter(grepl("Interferon|interferon", path_name))
INF_members <- pathwaymembers %>%
  filter(DB_ID %in% interferon$DB_ID) %>%
  distinct(gene_name)
YAP <- pathway_names %>%
  filter(grepl("YAP|TEAD", path_name))
YAP_members <- pathwaymembers %>%
  filter(DB_ID %in% YAP$DB_ID) %>%
  distinct(gene_name)
cell_cycle <- pathway_names %>%
  filter(grepl("Cell cycle|cell cycle|APC|mitotic", path_name))
cc_members <- pathwaymembers %>%
  filter(DB_ID %in% cell_cycle$DB_ID) %>%
  distinct(gene_name)
AP2 <- pathway_names %>%
  filter(grepl("TFAP", path_name))
AP2_members <- pathwaymembers %>%
  filter(DB_ID %in% AP2$DB_ID) %>%
  distinct(gene_name)
```

# Load and process networks 


```{r, eval = F}
# noTPP
all_input_nodes <- c(viper_TFs$id, viper_kinases$id)
noTPP <- c()

load("results/networks/2021_08_22_10_04_11_noTPP_PKN1_rep1.RData")
res_noTPP_PKN1_rep1 = network_processing(COSMOS_object_fw2, COSMOS_object_rev2, noTPP, dorothea, KSN, all_input_nodes)
netprop_noTPP_PKN1_rep1 <- network_checks(res_noTPP_PKN1_rep1, noTPP, viper_TFs$id, viper_kinases$id, dorothea, KSN)

load("results/networks/2021_08_22_21_30_33_noTPP_PKN1_rep2.RData")
res_noTPP_PKN1_rep2 = network_processing(COSMOS_object_fw2, COSMOS_object_rev2, noTPP, dorothea, KSN, all_input_nodes)
netprop_noTPP_PKN1_rep2 <- network_checks(res_noTPP_PKN1_rep2, noTPP, viper_TFs$id, viper_kinases$id, dorothea, KSN)

load("results/networks/2021_08_23_07_36_48_noTPP_PKN1_rep3.RData")
res_noTPP_PKN1_rep3 = network_processing(COSMOS_object_fw2, COSMOS_object_rev2, noTPP, dorothea, KSN, all_input_nodes)
netprop_noTPP_PKN1_rep3 <- network_checks(res_noTPP_PKN1_rep3, noTPP, viper_TFs$id, viper_kinases$id, dorothea, KSN)

## PKN2

load("results/networks/2021_09_09_12_03_31_noTPP_PKN2_rep1.RData")
res_noTPP_PKN2_rep1 = network_processing(COSMOS_object_fw2, COSMOS_object_rev2, noTPP, dorothea, KSN, all_input_nodes)
netprop_noTPP_PKN2_rep1 <- network_checks(res_noTPP_PKN2_rep1, noTPP, viper_TFs$id, viper_kinases$id, dorothea, KSN)

load("results/networks/2021_09_09_15_08_40_noTPP_PKN2_rep2.RData")
res_noTPP_PKN2_rep2 = network_processing(COSMOS_object_fw2, COSMOS_object_rev2, noTPP, dorothea, KSN, all_input_nodes)
netprop_noTPP_PKN2_rep2 <- network_checks(res_noTPP_PKN2_rep2, noTPP, viper_TFs$id, viper_kinases$id, dorothea, KSN)


load("results/networks/2021_09_09_17_51_40_noTPP_PKN2_rep3.RData")
res_noTPP_PKN2_rep3 = network_processing(COSMOS_object_fw2, COSMOS_object_rev2, noTPP, dorothea, KSN, all_input_nodes)
netprop_noTPP_PKN2_rep3 <- network_checks(res_noTPP_PKN2_rep3, noTPP, viper_TFs$id, viper_kinases$id, dorothea, KSN)

# withTPP
all_input_nodes <- c(viper_TFs$id, viper_kinases$id, TPP_cosmos_input_final$TPP_protein)
TPP <- TPP_cosmos_input_final$TPP_protein

load("results/networks/2021_08_26_10_59_55_withTPP_PKN1_rep1.RData")
res_withTPP_PKN1_rep1 = network_processing(COSMOS_object_fw2, COSMOS_object_rev2, TPP, dorothea, KSN, all_input_nodes)
netprop_withTPP_PKN1_rep1 <- network_checks(res_withTPP_PKN1_rep1, TPP, viper_TFs$id, viper_kinases$id, dorothea, KSN)

load("results/networks/2021_08_26_13_46_35_withTPP_PKN1_rep2.RData")
res_withTPP_PKN1_rep2 = network_processing(COSMOS_object_fw2, COSMOS_object_rev2, TPP, dorothea, KSN, all_input_nodes)
netprop_withTPP_PKN1_rep2 <- network_checks(res_withTPP_PKN1_rep2, TPP, viper_TFs$id, viper_kinases$id, dorothea, KSN)


load("results/networks/2021_08_26_15_29_33_withTPP_PKN1_rep3.RData")
res_withTPP_PKN1_rep3 = network_processing(COSMOS_object_fw2, COSMOS_object_rev2, TPP, dorothea, KSN, all_input_nodes)
netprop_withTPP_PKN1_rep3 <- network_checks(res_withTPP_PKN1_rep3, TPP, viper_TFs$id, viper_kinases$id, dorothea, KSN)

## PKN2

load("results/networks/2021_09_05_12_15_40_withTPP_PKN2_rep1.RData")
res_withTPP_PKN2_rep1 = network_processing(COSMOS_object_fw2, COSMOS_object_rev2, TPP, dorothea, KSN, all_input_nodes)
netprop_withTPP_PKN2_rep1 <- network_checks(res_withTPP_PKN2_rep1, TPP, viper_TFs$id, viper_kinases$id, dorothea, KSN)

load("results/networks/2021_09_05_18_59_21_withTPP_PKN2_rep2.RData")
res_withTPP_PKN2_rep2 = network_processing(COSMOS_object_fw2, COSMOS_object_rev2, TPP, dorothea, KSN, all_input_nodes)
netprop_withTPP_PKN2_rep2 <- network_checks(res_withTPP_PKN2_rep2, TPP, viper_TFs$id, viper_kinases$id, dorothea, KSN)


load("results/networks/2021_09_06_01_41_33_withTPP_PKN2_rep3.RData")
res_withTPP_PKN2_rep3 = network_processing(COSMOS_object_fw2, COSMOS_object_rev2, TPP, dorothea, KSN, all_input_nodes)
netprop_withTPP_PKN2_rep3 <- network_checks(res_withTPP_PKN2_rep3, TPP, viper_TFs$id, viper_kinases$id, dorothea, KSN)

rm(COSMOS_object_fw1, COSMOS_object_fw2, COSMOS_object_rev1, COSMOS_object_rev2)
```

```{r, eval = F}
result_list <- mget(ls(pattern= "^res_"))
names(result_list) <- c("res_noTPP_PKN1_rep1","res_noTPP_PKN1_rep2","res_noTPP_PKN1_rep3",
                        "res_noTPP_PKN2_rep1","res_noTPP_PKN2_rep2",   "res_noTPP_PKN2_rep3",
                        "res_withTPP_PKN1_rep1","res_withTPP_PKN1_rep2", "res_withTPP_PKN1_rep3", 
                        "res_withTPP_PKN2_rep1", "res_withTPP_PKN2_rep2", "res_withTPP_PKN2_rep3") 

network_list <- mget(ls(pattern= "^netprop_"))
names(network_list) <- c("netprop_noTPP_PKN1_rep1","netprop_noTPP_PKN1_rep2","netprop_noTPP_PKN1_rep3",
                         "netprop_noTPP_PKN2_rep1","netprop_noTPP_PKN2_rep2","netprop_noTPP_PKN2_rep3",  
                         "netprop_withTPP_PKN1_rep1", "netprop_withTPP_PKN1_rep2", "netprop_withTPP_PKN1_rep3",
                         "netprop_withTPP_PKN2_rep1", "netprop_withTPP_PKN2_rep2", "netprop_withTPP_PKN2_rep3") 

rm(list = ls(pattern= "^res_"))
rm(list = ls(pattern= "^netprop_"))
```


## Save collapsed network objects

```{r, eval = F}
save(result_list, network_list, file = "results/networks/220531_mergednetworkresults.RData")
```

```{r}
load("results/networks/220531_mergednetworkresults.RData")
```


```{r}
#load("data/2022_02_17_MasterThesis_networks.RData")
```

## Format networks

```{r}
nodes <- map_dfr(result_list, "nodes", .id = "networkID") %>%
  mutate(networkID = gsub("^res_", "", networkID)) %>% 
  dplyr::select(networkID, label, measured)

nodes <- map_dfr(network_list, "overall_node_summary", .id = "networkID") %>%
  mutate(networkID = gsub("^netprop_", "", networkID)) %>% 
  separate(networkID, into = c("TPP_set", "PKN", "replicate"), sep = "_", remove = F) %>%
  left_join(nodes, by = c("networkID", "node" = "label")) %>%
  mutate(
    model = gsub("input", "input node modelled", model),
    model = gsub("missed", "input node not modelled", model),
    model = gsub("additional", "PKN node", model),
    id = node
  )

edges <- map_dfr(result_list, "combined_network", .id = "networkID") %>%
  mutate(networkID = gsub("^res_", "", networkID)) %>% 
  mutate(edge_combined = paste(from, sign, to, sep = "_")) %>%
  distinct(networkID, edge_combined, .keep_all = T) %>%
  separate(networkID, into = c("TPP_set", "PKN", "replicate"), sep = "_", remove = F)
```

Extract F-statistic results and results without TPP

```{r}
withTPP_edges <- filter(edges, TPP_set == "withTPP") %>%
  distinct(edge_combined, .keep_all = T)

withTPP_nodes <- data.frame(id = unique(c(withTPP_edges$from, withTPP_edges$to))) %>%
  left_join(distinct(nodes, id, .keep_all = T)) %>%
  mutate(group = paste(type, measured, sep = "_"))

withTPP_nodes <- nodes %>%
  filter(TPP_set == "withTPP" & model != "input node not modelled" & measured != "not available") %>%
  distinct(id, .keep_all = T) %>%
  mutate(group = paste(type, measured, sep = "_"))

# annotate with footprinting information 
withTPP_nodes <- bind_rows(
  filter(viper_kinases, id %in% withTPP_nodes$id) %>% mutate(model = "input node modelled"),
  withTPP_nodes
) %>%
  distinct(id, .keep_all = T) %>%
  mutate(group = paste(type, measured, sep = "_"))

noTPP_edges <- filter(edges, TPP_set == "noTPP") %>%
  distinct(edge_combined, .keep_all = T)

noTPP_nodes <- data.frame(id = unique(c(noTPP_edges$from, noTPP_edges$to))) %>%
  left_join(distinct(nodes, id, .keep_all = T)) %>%
  mutate(group = paste(type, measured, sep = "_"))

noTPP_nodes <- nodes %>%
  filter(TPP_set == "noTPP" & model != "input node not modelled" & measured != "not available") %>%
  distinct(id, .keep_all = T) %>%
  mutate(group = paste(type, measured, sep = "_"))
```

# Enrichment

```{r, message=F, warning=F}
# universe built from PKN
pathways_universe <- data.frame(protein = PKN_universe$Node, NES = 1)

# get gathered pathway input
pathway_input_gathered <- map_dfr(result_list, 'nodes', .id = "networkID") %>% 
  mutate(networkID = gsub("^res_", "", networkID)) %>%
  dplyr::select(networkID, id) %>% 
  filter(grepl("noTPP|withTPP", networkID))

gene_lists <- pathway_input_gathered %>% 
  group_by(networkID) %>% 
  summarise(length = n())

# Reactome enrichment
pathway_enrichment <- pathway_input_gathered %>% 
  group_by(networkID) %>% 
  mutate(enrichment = list(as.data.frame(GSE_RPA(geneList = id, universe = pathways_universe$protein))))

pathway_enrichment_unnest <- pathway_enrichment %>% 
  distinct(networkID, enrichment) %>% 
  unnest() %>% 
  separate(networkID, into = c("TPP_set", "PKN", "replicate"), sep = "_", remove = F) %>% 
  left_join(gene_lists, by = "networkID")
```

Calculate log odds ratio

```{r, message=F, warning=F}
pathway_enrichment_significant <- pathway_enrichment_unnest %>% 
  filter(qvalue < 0.05) %>% 
  # add some variables
  mutate(pathway = Description, Zscore = -log10(qvalue),
         pathway_size = as.numeric(gsub("/.+$", "", BgRatio)),
         Gene_Ratio = Count/pathway_size,
         LR = log2((Count/length)/(pathway_size/length(pathways_universe$protein))))
```


Disentangle the results

```{r}
pathways_perstrategy <- pathway_enrichment_significant %>% 
  ungroup() %>% 
  group_by(PKN, TPP_set, pathway) %>% 
  mutate(n_replicates = n()) %>% 
  # filter pathways to be found i  at least two of three network replicates
  filter(n_replicates > 1) %>% 
  # collapse replicate infromation
  distinct(PKN, TPP_set, pathway, .keep_all = T) %>% 
  select(TPP_set, PKN, c(15:19), qvalue, Count, geneID)

pathways_TPPunique <- pathways_perstrategy %>% 
  # use only
   anti_join(filter(pathway_enrichment_significant, TPP_set == "noTPP" ), by = c("PKN","pathway"))

pathways_noTPP <- pathways_perstrategy %>% 
 filter(TPP_set == "noTPP" )
```

### Compare enrichment

```{r, fig.width=4, fig.height=4}

group_cc <- "cell cycle|Cell Cycle|NIMA|APC|G2/|G1/|Mitotic|mitotic|Prometaphase|G2 Phase|G1 Phase|Prophase|cyclin|Cyclin"
group_DNA <- "Checkpoint|Repair|Damage|repair|damage| Strand Breaks|DNA double-strand break|Homologous Recombination|Chk1/Chk2|DNA Double Strand Break"

pathway_grouping <- function(ER_result, group, exclude=F){
  grouped_pathways <- ER_result %>%
  ungroup() %>%
  distinct(PKN, pathway, .keep_all = T) %>%
  select(PKN, pathway, qvalue, geneID, Count, Gene_Ratio, pathway_size, LR)
  if(exclude ==F){
    grouped_pathways <- grouped_pathways %>% filter(grepl(group, pathway))
  } else {
    grouped_pathways <- grouped_pathways %>% filter(!grepl(group, pathway))
  }
  
}

TPP_cc_pathways <- pathway_grouping(pathways_TPPunique, group_cc)
TPP_DDR_pathways <- pathway_grouping(pathways_TPPunique, group_DNA)
TPP_other_pathways <- pathway_grouping(pathways_TPPunique, paste(group_cc, group_DNA, sep ="|"), exclude =T)


comparison_TPP <- bind_rows(
  "Cell \n Cycle" = TPP_cc_pathways,
  "DNA damage \n and repair" = TPP_DDR_pathways,
  "other" = TPP_other_pathways,
  .id = "pathway_family")

noTPP_cc_pathways <- pathway_grouping(pathways_noTPP, group_cc)
noTPP_DDR_pathways <- pathway_grouping(pathways_noTPP, group_DNA)
noTPP_other_pathways <- pathway_grouping(pathways_noTPP, paste(group_cc, group_DNA, sep ="|"), exclude =T)

comparison_noTPP <- bind_rows(
  "Cell \n Cycle" = noTPP_cc_pathways,
  "DNA damage \n and repair" = noTPP_DDR_pathways,
  "other" = noTPP_other_pathways,
  .id = "pathway_family") 

comparison = bind_rows("TPP" = comparison_TPP,
                       "noTPP" = comparison_noTPP,
                       .id = "network") %>% 
  # filter for minimum two nodes matched to a pathway with minimum 6 annotated members
  filter(Count > 1 & pathway_size > 5) %>% 
  ungroup()

p <- ggplot(comparison, aes(x = pathway_family, y = LR, colour = network)) +
  geom_boxplot(lwd =1) +
  stat_compare_means(aes(group = network), label.y = 7, label = "p.signif") +
  scale_colour_manual(values = c("TPP" = "mediumseagreen", "noTPP" = "grey50")) +
  cowplot::theme_cowplot(font_size = 16) +
  labs(x = "", y = "Log odds-ratio \n pathway enrichment") +
  theme(legend.position = "bottom")

p

```



## Network plots

```{r}
set.seed(1)
p <- visNetwork(withTPP_nodes, withTPP_edges) %>%
  visGroups(groupname = "TPP_active", color = "mediumseagreen") %>%
  visGroups(groupname = "TPP_inactive", color = "palegreen") %>%
  visGroups(groupname = "TF_active", color = "#8b0a50") %>%
  visGroups(groupname = "TF_inactive", color = "palevioletred") %>%
  visGroups(groupname = "Kinase_active", color = "darkcyan") %>%
  visGroups(groupname = "Kinase_inactive", color = "mediumturquoise") %>%
  visGroups(groupname = "other_active", color = "burlywood") %>%
  visGroups(groupname = "other_inactive", color = "wheat") %>%
  # standard options
  visEdges(arrows = "to", color = "dimgrey", width = 2) %>%
  visNodes(font = "30px arial black bold") %>%
  visLegend() %>%
  visOptions(highlightNearest = TRUE, nodesIdSelection = TRUE) %>%
  visNetwork::visIgraphLayout(layout = "layout_with_kk")
p
```


```{r}
set.seed(1)
p <- visNetwork(noTPP_nodes, noTPP_edges) %>%
  visGroups(groupname = "TPP_active", color = "mediumseagreen") %>%
  visGroups(groupname = "TPP_inactive", color = "palegreen") %>%
  visGroups(groupname = "TF_active", color = "#8b0a50") %>%
  visGroups(groupname = "TF_inactive", color = "palevioletred") %>%
  visGroups(groupname = "Kinase_active", color = "darkcyan") %>%
  visGroups(groupname = "Kinase_inactive", color = "mediumturquoise") %>%
  visGroups(groupname = "other_active", color = "burlywood") %>%
  visGroups(groupname = "other_inactive", color = "wheat") %>%
  # standard options
  visEdges(arrows = "to", color = "dimgrey", width = 2) %>%
  visNodes(font = "30px arial black bold") %>%
  visLegend() %>%
  visOptions(highlightNearest = TRUE, nodesIdSelection = TRUE) %>%
  visNetwork::visIgraphLayout(layout = "layout_with_kk")
p
```


# F statistic cluster analysis

Cluster network and annotate nodes with cluster information

```{r}
graph_withTPP <- withTPP_edges %>%
  select(from, to, sign) %>%
  distinct(from, to, .keep_all = T) %>%
  graph_from_data_frame()

graph_UD <- as.undirected(graph_withTPP, mode = "collapse")
cfg <- igraph::cluster_fast_greedy(graph_UD)

n_clusters_withTPP <- length(unique(cfg$membership))
n_clusters_withTPP

clusters_withTPP <- data.frame(node = cfg$names, cluster = cfg$membership)

withTPP_nodes <- withTPP_nodes %>%
  left_join(clusters_withTPP, by = "node") %>%
  mutate(cluster = as.factor(cluster))
```

```{r , fig.width=13, fig.height=7}
pal <- c(
  "1" = "#b68c47",
  "2" = "#934ac0",
  "3" = "#71b448",
  "4" = "#af5179",
  "5" = "#558965",
  "6" = "#bb4b37",
  "7" = "#6d7ab3"
)

pal <- c(
  "1" = "#EE7733",
  "2" = "#0077BB",
  "3" = "#33BBEE",
  "4" = "#EE3377",
  "5" = "#CC3311",
  "6" = "#009988",
  "7" = "#BBBBBB"
)

target <- clusters_withTPP$node
withTPP_nodes <- withTPP_nodes[match(target, withTPP_nodes$id), ] %>% 
  mutate(shape = factor(withTPP_nodes$cluster))

set.seed(4)
p <- ggraph(graph_withTPP, layout = "fr") +
  geom_edge_link(
    colour = "grey50",
    end_cap = circle(3, "mm"),
    start_cap = circle(3, "mm")
  ) +
  geom_node_point(aes(colour = withTPP_nodes$cluster), size = 6) +
  scale_colour_manual(values = pal) +
 # scale_shape_manual(values = c(15:20, 25)) +
  theme_graph() +
  theme(legend.position = "right")
p
```



## Cluster level enrichment

```{r}
# PKN as universe
universe <- PKN_universe$Node

pathwayenrichment_cluster <- withTPP_nodes %>%
  group_by(cluster) %>%
  mutate(enrichment = list(as.data.frame(GSE_RPA(
    geneList = id,
    universe = universe
  ))))

pathway_enrichment_unnest_withTPP <- pathwayenrichment_cluster %>%
  distinct(cluster, enrichment) %>%
  unnest()
```

Extract top pathways per cluster

```{r}
pathway_enrichment_top3cluster <- pathway_enrichment_unnest_withTPP %>%
  filter(Count > 2) %>%
  ungroup() %>%
  mutate(corr_p_value = -log10(p.adjust)) %>%
  group_by(cluster) %>%
  dplyr::arrange(cluster, desc(corr_p_value), desc(Count)) %>%
  slice_head(n = 3) %>%
  # rip apart labels for better visualization
  mutate(id = letters[row_number()], clusterID = paste(cluster, id, sep = "_")) %>%
  ungroup() %>%
  mutate(cluster = as.factor(cluster), label = gsub("(.{25})", "\\1-\n", Description))

pathway_nodes <- data.frame(
  from = pathway_enrichment_top3cluster$cluster,
  to = pathway_enrichment_top3cluster$clusterID
)

# simplify graph to cluster level information
V(graph_withTPP)$name <- withTPP_nodes$cluster
g_cont <- contract(graph_withTPP, membership(cfg), vertex.attr.comb = toString)

g_simp <- as_data_frame(igraph::simplify(g_cont)) %>%
  mutate(from = gsub(",.*$", "", from), to = gsub(",.*$", "", to)) %>%
  bind_rows(pathway_nodes) %>%
  graph_from_data_frame()
```

```{r}
pathway_enrichment_topcluster <- pathway_enrichment_unnest_withTPP %>%
  dplyr::ungroup() %>%
  dplyr::mutate(corr_p_value = -log10(p.adjust)) %>%
  dplyr::group_by(cluster) %>%
  dplyr::arrange(cluster, desc(corr_p_value), desc(Count)) %>%
  dplyr::slice_head(n = 1)

pathway_enrichment_topcluster %>% select(cluster, Description, qvalue)
```
## Processes in detail 

Analyze the processes identified by the cluster level analysis in detail.

```{r}
for (i in list(
  AP2_members$gene_name,
  YAP_members$gene_name,
  INF_members$gene_name,
  c(DDR_members$gene_name, cc_members$gene_name)
)) {
  
  interesting_nodes <- withTPP_nodes %>%
    filter(id %in% i)

  interesting_edges <- withTPP_edges %>%
    filter(from %in% interesting_nodes$id | to %in% interesting_nodes$id) %>% 
    mutate(color = ifelse(from %in% interesting_nodes$id & to %in% interesting_nodes$id, "grey", "grey")) %>% 
    distinct(from, to, .keep_all=T)

  occurences <- table(c(interesting_edges$from, interesting_edges$to))

  interesting_nodes <- withTPP_nodes %>%
    filter(id %in% unique(c(interesting_edges$from, interesting_edges$to))) %>%
    mutate(
      group = ifelse(model == "PKN node", "related", "pathway"),
      size = ifelse(type == "TPP", 35, 20),
      # bigger nodes for pathway nodes
      shape = ifelse(type == "TPP", "star", "dot"),
      label = str_pad(node, max(nchar(node)), side="both", pad=" ")
    ) %>%
    rowwise() %>%
    # make network extract smaller and better readable by reomving nodes with only one edge
    filter(id %in% i | occurences[[id]] > 1)

  set.seed(1)
  p <- visNetwork(interesting_nodes, interesting_edges) %>%
     visGroups(groupname = "pathway", color = "darkred") %>%
      visGroups(groupname = "related", color = "grey") %>%
    # visGroups(groupname = "TF_active", color = "#8b0a50") %>%
    # visGroups(groupname = "TF_inactive", color = "palevioletred") %>%
    # visGroups(groupname = "Kinase_active", color = "darkcyan") %>%
    # visGroups(groupname = "Kinase_inactive", color = "mediumturquoise") %>%
    # visGroups(groupname = "other_active", color = "burlywood") %>%
    # visGroups(groupname = "other_inactive", color = "wheat") %>%
    # visGroups(groupname = "background", color = "gainsboro") %>%
    # standard options
    visEdges(arrows = "to", width = 4) %>%
    visNodes(font = "40px arial black bold") %>%
    visLegend() %>%
    visOptions(highlightNearest = TRUE, nodesIdSelection = TRUE) %>%
    visNetwork::visIgraphLayout(layout = "layout_with_fr")
  print(p)
}
```




# noTPP cluster analysis

Cluster network and annotate nodes with cluster information

```{r}
graph_noTPP <- noTPP_edges %>%
  select(from, to, sign) %>%
  distinct(from, to, .keep_all = T) %>%
  graph_from_data_frame()

graph_UD <- as.undirected(graph_noTPP, mode = "collapse")
cfg <- igraph::cluster_fast_greedy(graph_UD)

n_clusters_noTPP <- length(unique(cfg$membership))
n_clusters_noTPP

clusters_noTPP <- data.frame(node = cfg$names, cluster = cfg$membership)

noTPP_nodes <- noTPP_nodes %>%
  left_join(clusters_noTPP, by = "node") %>%
  mutate(cluster = as.factor(cluster))
```

```{r , fig.width=15, fig.height=10}
pal <- c(
  "1" = "#b68c47",
  "2" = "#934ac0",
  "3" = "#71b448",
  "4" = "#af5179",
  "5" = "#558965",
  "6" = "#bb4b37",
  "7" = "#6d7ab3",
  "8" = "darkgrey"
)

target <- clusters_noTPP$node
noTPP_nodes <- noTPP_nodes[match(target, noTPP_nodes$id), ]

set.seed(4)
p <- ggraph(graph_noTPP, layout = "fr") +
  geom_edge_link(
    colour = "grey50",
    end_cap = circle(3, "mm"),
    start_cap = circle(3, "mm")
  ) +
  geom_node_point(aes(colour = noTPP_nodes$cluster), size = 6) +
  geom_node_text(aes(label = noTPP_nodes$id), size = 2.5) +
  scale_colour_manual(values = pal) +
  theme_graph() +
  theme(legend.position = "right")
p
```

## Cluster level enrichment

Check how clusters of noTPP networks compare to the insights in networks with TPP.

```{r}
# PKN as universe
universe <-PKN_universe$Node

pathwayenrichment_cluster <- noTPP_nodes %>%
  group_by(cluster) %>%
  mutate(enrichment = list(as.data.frame(GSE_RPA(
    geneList = id,
    universe = universe
  ))))

pathway_enrichment_unnest_noTPP <- pathwayenrichment_cluster %>%
  distinct(cluster, enrichment) %>%
  unnest()
```

Extract top pathways per cluster

```{r}
pathway_enrichment_top3cluster <- pathway_enrichment_unnest_noTPP %>%
  filter(Count > 2) %>%
  ungroup() %>%
  mutate(corr_p_value = -log10(p.adjust)) %>%
  group_by(cluster) %>%
  dplyr::arrange(cluster, desc(corr_p_value), desc(Count)) %>%
  slice_head(n = 3) %>%
  # rip apart labels for better visualization
  mutate(id = letters[row_number()], clusterID = paste(cluster, id, sep = "_")) %>%
  ungroup() %>%
  mutate(cluster = as.factor(cluster), label = gsub("(.{25})", "\\1-\n", Description))

pathway_nodes <- data.frame(
  from = pathway_enrichment_top3cluster$cluster,
  to = pathway_enrichment_top3cluster$clusterID
)

# simplify graph to cluster level information
V(graph_noTPP)$name <- noTPP_nodes$cluster
g_cont <- contract(graph_noTPP, membership(cfg), vertex.attr.comb = toString)

g_simp <- as_data_frame(igraph::simplify(g_cont)) %>%
  mutate(from = gsub(",.*$", "", from), to = gsub(",.*$", "", to)) %>%
  bind_rows(pathway_nodes) %>%
  graph_from_data_frame()
```

```{r}
pathway_enrichment_topcluster <- pathway_enrichment_unnest_noTPP %>%
  ungroup() %>%
  mutate(corr_p_value = -log10(p.adjust)) %>%
  group_by(cluster) %>%
  dplyr::arrange(cluster, desc(corr_p_value), desc(Count)) %>%
  slice_head(n = 1)

pathway_enrichment_topcluster %>% select(cluster, Description, qvalue)
```

# Compare clusters

```{r}
pathway_enrichment_topclusters_withTPP <- pathway_enrichment_unnest_withTPP %>%
  ungroup() %>%
  mutate(corr_p_value = -log10(p.adjust)) %>%
  group_by(cluster) %>%
  dplyr::arrange(cluster, desc(corr_p_value), desc(Count)) %>%
  slice_head(n = 5)

pathway_enrichment_topclusters_noTPP <- pathway_enrichment_unnest_noTPP %>%
  ungroup() %>%
  mutate(corr_p_value = -log10(p.adjust)) %>%
  group_by(cluster) %>%
  dplyr::arrange(cluster, desc(corr_p_value), desc(Count)) %>%
  slice_head(n = 5)

intersect(pathway_enrichment_topclusters_withTPP$Description, pathway_enrichment_topclusters_noTPP$Description)

merged_clusters <- pathway_enrichment_topclusters_withTPP %>%
  select(cluster, Description) %>%
  mutate(network = "with TPP") %>%
  bind_rows(pathway_enrichment_topclusters_noTPP %>%
    select(cluster, Description) %>%
    mutate(network = "no TPP"))
```

## Venn diagram

```{r}
input_for_venn <- merged_clusters %>%
  dcast(Description ~ network) %>%
  mutate_at(.vars = vars(c(2:3)), as.logical)
```


```{r}
p_venn1 <- ggplot(input_for_venn) +
  geom_venn(aes(
    A = `with TPP`,
    B = `no TPP`
  ),
  fill_color = c("mediumseagreen", "grey"),
  fill_alpha = 0.8,
  stroke_alpha = 0, text_size = 10,
  set_name_size = 10,
  show_percentage = FALSE
  ) +
  coord_fixed() +
  theme_void()
p_venn1
```



## Jaccard index analysis


```{r}
merged_clusters_comparisons <- withTPP_nodes %>%
  select(node, cluster) %>%
  mutate(network = "with TPP") %>%
  bind_rows(noTPP_nodes %>% select(node, cluster) %>% mutate(network = "no TPP")) %>%
  ungroup() %>%
  mutate(id = row_number()) %>%
  acast(id ~ network + cluster, value.var = "node")

LoL <- apply(merged_clusters_comparisons, 2, as.list)
LoL <- lapply(LoL, function(x) x[!is.na(x)])

overlaps <- sapply(LoL, function(x) {
  sapply(LoL, function(y, x) length(intersect(x, y)), x)
}) %>%
  melt() %>%
  setNames(c("cond1", "cond2", "overlap"))

Jaccard_index <- sapply(LoL, function(x) {
  sapply(LoL, function(y, x) length(intersect(x, y)) / length(union(x, y)), x)
}) %>%
  melt() %>%
  setNames(c("cond1", "cond2", "JI")) %>%
  left_join(overlaps)
```



```{r}
pal <- c(
  "1" = "#EE7733",
  "2" = "#0077BB",
  "3" = "#33BBEE",
  "4" = "#EE3377",
  "5" = "#CC3311",
  "6" = "#009988",
  "7" = "#BBBBBB",
  "no overlap" = "white"
)

bb <- c(0,0.25, 0.5, 0.75, 1) # define breaks.
#ll <- c("40%","30%","20%","10%","5%") # labels.


p_JI <- Jaccard_index %>%
  filter(cond1 != cond2) %>%
  filter(!(grepl("with", cond1) & grepl("with", cond2))) %>%
  filter(!(grepl("no", cond1) & grepl("no", cond2))) %>%
  dplyr::mutate(normalized = purrr::map2_chr(cond1, cond2, ~ paste(sort(c(.x, .y)), collapse = ""))) %>%
  distinct(normalized, .keep_all = T) %>%
  dplyr::select(-normalized) %>%
  mutate(
    cluster_noTPP = gsub("no TPP_", "", cond2), cluster_withTPP = gsub("with TPP_", "", cond1),
    colouring = ifelse(JI == 0, "no overlap", cluster_withTPP),
    rounded = ifelse(JI != 0, round(JI,2), NA)
  ) %>%
  ggplot(aes(x = (cluster_withTPP), y = (cluster_noTPP), colour = colouring, size = JI)) +
  geom_point() +
  scale_size_continuous(name = "JI",
                          breaks = bb,
                          limits = c(0, 1),
                          range = c(1, 10) ) +
  scale_colour_manual(values = pal, guide = "none") +
  cowplot::theme_cowplot(font_size = 16) +
  theme(panel.grid = element_blank(), legend.position = "bottom") +
  labs(x = "clusters networks with TPP", y = "clusters networks without TPP")
p_JI
```


# Save
```{r, eval =FALSE}

```
